requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-dotenv==1.0.0
groq==0.9.0
PyPDF2==3.0.1
scikit-learn>=1.4.0
numpy>=1.26
gunicorn==21.2.0
whitenoise==6.6.0
